refactor(agents): extract termination seam from agents.ts
Move stopAgentSync, stopAgent, killLauncherProcessSync, killLauncherProcessAsync,
and isTmuxProcessSync to src/lib/agents/termination.ts; re-export from agents.ts.

Behavior-preserving move only; no logic changes.
